split again (this time correctly)
