+++ src/st/np3/3DEF4.c
@@ -1094,723 +1094,6 @@ INCLUDE_ASM("asm/us/st/np3/nonmatchings/3DEF4", func_801C5F58);
 // EntityWaterDrop ID 0x31
 INCLUDE_ASM("asm/us/st/np3/nonmatchings/3DEF4", func_801C61B4);
 
-s32 func_801C6458(s16 yVector) {
-    s16 newY = yVector + g_CurrentEntity->posY.i.hi;
-    s32 expectedResult = 0;
-    Collider collider;
-    Entity* newEntity;
-    s32 res;
 
-    g_api.CheckCollision(g_CurrentEntity->posX.i.hi, newY, &collider, 0);
-    res = expectedResult == (collider.unk0 & 1);
-
-    if (collider.unk0 & 8) {
-        if (*(u8*)&g_CurrentEntity->ext.generic.unkA0 == 0) {
-            newEntity = AllocEntity(&D_8007DE38, &D_8007DE38[24]);
-            if (newEntity != NULL) {
-                CreateEntityFromEntity(0x33, g_CurrentEntity, newEntity);
-                newEntity->posY.i.hi += yVector;
-                newEntity->zPriority = g_CurrentEntity->zPriority;
-            }
-            g_api.PlaySfx(0x7C2);
-            *(u8*)&g_CurrentEntity->ext.generic.unkA0 = 1;
-        }
-    }
-    return res;
-}
-
-#ifndef NON_MATCHING
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3DEF4", EntityMerman2);
-#else
-#define NA_SE_EV_WATER_SPLASH 0x7C2
-
-extern u16 D_80181230[]; // pos TBL
-extern s32 D_80182224[]; // accel_x TBL
-extern s32 D_80182228[]; // accel_y TBL
-extern u8 D_80182244[];  // timers
-extern s32 D_80182248;   // Collision data
-extern s32 D_80182258;   // Collision data
-extern u16 D_80182268;   // Collision data
-extern u8 D_80182270[];  // Animation
-extern u8 D_8018227C[];  // Animation
-extern u8 D_80182294[];  // Animation
-extern u8 D_8018229C[];  // Animation
-
-void EntityMerman2(Entity* self) {
-    s32 firstPrimIndex;
-    Collider collider;
-    Entity* newEntity;
-    Entity* temp_s4;
-    Primitive* prim;
-    s32 colRes;
-    s32 offset;
-    s16* pos;
-    s16 posX;
-    s16 posY;
-    s16 camY;
-    s32 rnd;
-    s32 i;
-    u16 temp_v0_10;
-    u16 temp_v0_5;
-    u16 temp_v0_6;
-    u16 temp_v0_9;
-    u32 temp_v0_12;
-    u32 v0;
-    u16 var_v0_11;
-    s32 res;
-    s32 index;
-    unsigned int temp;
-    if (self->ext.merman2.ignoreCol && (self->step < 7)) {
-        SetStep(7);
-    }
-    if ((self->flags & 0x100) && (self->step < 8)) {
-        func_801C2598(0x71D);
-        self->unk19 = 0;
-        if (self->flags & 0x00800000) {
-            g_api.FreePrimitives(self->firstPolygonIndex);
-            self->flags &= ~0x00800000;
-        }
-        self->unk3C = 0;
-        self->flags &= ~0x20000000;
-        SetStep(8);
-    }
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180AB4);
-        self->unk12 = 8;
-        self->zPriority = 0xA9;
-        if (self->subId != 0) {
-            self->palette++;
-        }
-        self->accelerationY = -0x10000;
-        break;
-
-    case 1:
-        AnimateEntity(D_8018229C, self);
-        MoveEntity();
-        posY = self->posY.i.hi - 24;
-        g_api.CheckCollision(self->posX.i.hi, posY, &collider, 0);
-        if (!(collider.unk0 & 8)) {
-            SetStep(3);
-            break;
-        }
-
-    case 3:
-        if (self->step_s == 0) {
-            rnd = (Random() & 3) * 2;
-            self->accelerationX = D_80182224[rnd];
-            self->accelerationY = D_80182228[rnd];
-            self->step_s++;
-        }
-        if (AnimateEntity(D_8018229C, self) == 0) {
-            self->facing = (GetPlayerSide() & 1) ^ 1;
-        }
-        MoveEntity();
-        posX = self->posX.i.hi;
-        posY = self->posY.i.hi;
-        posY -= 24;
-        g_api.CheckCollision(posX, posY, &collider, 0);
-        if (!(collider.unk0 & 8)) {
-            self->accelerationY = 0x8000;
-        }
-        index = (self->subId >> 8) & 1;
-        pos = &D_80181230;
-        pos += index;
-        posY += g_Camera.posY.i.hi;
-        if (pos[4] < posY) {
-            self->posY.i.hi = (pos[4] - g_Camera.posY.i.hi) - 24;
-        }
-        if (((u8)(self->ext.merman2.timer++)) > 32) {
-            self->ext.merman2.timer = 0;
-            self->step_s = 0;
-            if ((GetPlayerDistanceX() >= 48) && (!(Random() & 1))) {
-                SetStep(4);
-            }
-        }
-        break;
-
-    case 4:
-        switch (self->step_s) {
-        case 0:
-            self->accelerationX = 0;
-            self->accelerationY = -0x88000;
-            self->step_s++;
-            firstPrimIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
-            if (firstPrimIndex != (-1)) {
-                prim = &g_PrimBuf[firstPrimIndex];
-                self->animCurFrame = 0;
-                self->firstPolygonIndex = firstPrimIndex;
-                self->ext.merman2.prim = prim;
-                self->flags |= 0x00800000;
-                prim->tpage = 0x12;
-                prim->clut = 0x28C;
-                if (self->subId & 1) {
-                    prim->clut = 0x28D;
-                }
-                prim->u0 = 0;
-                prim->u1 = 0x20;
-                prim->v1 = (prim->v0 = 0x38);
-                prim->u2 = prim->u0;
-                prim->u3 = prim->u1;
-                v0 = prim->v0 + 0x38;
-                prim->v2 = v0;
-                prim->v3 = v0;
-                if (self->facing != 0) {
-                    temp_v0_5 = self->posX.i.hi;
-                    prim->x0 = temp_v0_5 + 0xF;
-                    prim->x1 = temp_v0_5 - 0x11;
-                    prim->x2 = prim->x0;
-                    prim->x3 = prim->x1;
-                } else {
-                    temp_v0_6 = self->posX.i.hi;
-                    prim->x0 = temp_v0_6 - 0xF;
-                    prim->x1 = temp_v0_6 + 0x11;
-                    prim->x2 = prim->x0;
-                    prim->x3 = prim->x1;
-                }
-                prim->y0 = (prim->y1 = self->posY.i.hi - 0x11);
-                prim->r0 = 64, prim->g0 = 64, prim->b0 = 160;
-                *((s32*)(&prim->r1)) = *((s32*)(&prim->r0));
-                *((s32*)(&prim->r2)) = *((s32*)(&prim->r0));
-                *((s32*)(&prim->r3)) = *((s32*)(&prim->r0));
-                prim->y3 = (prim->y2 = prim->y0 + 0x38);
-                prim->priority = self->zPriority;
-                prim->blendMode = 6;
-                return;
-            }
-            self->animCurFrame = 17;
-            DestroyEntity(self);
-            break;
-
-        case 1:
-            MoveEntity();
-            index = (self->subId >> 8) & 1;
-            pos = &D_80181230;
-            pos += index;
-            posY = self->posY.i.hi;
-            posY -= 24;
-            camY = g_Camera.posY.i.hi + posY;
-            if (camY < pos[3]) {
-                g_api.PlaySfx(0x7C2);
-                newEntity = AllocEntity(D_8007DE38, &D_8007DE38[24]);
-                if (newEntity != 0) {
-                    CreateEntityFromEntity(0x33, self, newEntity);
-                    newEntity->posY.i.hi -= 24;
-                    newEntity->zPriority = self->zPriority;
-                }
-                self->unk1E = 0;
-                self->unk19 |= 4;
-                self->step_s++;
-            }
-            break;
-
-        case 2:
-            prim = self->ext.merman2.prim;
-            if (self->accelerationY > ((s32)0xFFFF4000)) {
-                prim->blendMode = 8;
-                self->animCurFrame = 18;
-                self->hitboxHeight = 8;
-                self->unk1E -= 0x80;
-            } else {
-                if (self->facing != 0) {
-                    temp_v0_9 = self->posX.i.hi;
-                    prim->x0 = temp_v0_9 + 0xF;
-                    prim->x1 = temp_v0_9 - 0x11;
-                    prim->x2 = prim->x0;
-                    prim->x3 = prim->x1;
-                    prim->y1 = (prim->y0 = self->posY.i.hi - 0x11);
-                    prim->y3 = (prim->y2 = prim->y0 + 0x38);
-                } else {
-                    temp_v0_10 = self->posX.i.hi;
-                    prim->x0 = temp_v0_10 - 0xF;
-                    prim->x1 = temp_v0_10 + 0x11;
-                    prim->x2 = prim->x0;
-                    prim->x3 = prim->x1;
-                    prim->y1 = (prim->y0 = self->posY.i.hi - 0x11);
-                    prim->y3 = (prim->y2 = prim->y0 + 0x38);
-                }
-                temp_v0_12 = ((u32)(((((self->accelerationY < 0)
-                                           ? (-self->accelerationY)
-                                           : (self->accelerationY)) >>
-                                      0xC) -
-                                     0xA) &
-                                    0xFF)) >>
-                             1;
-                prim->r0 = 128 - temp_v0_12, prim->g0 = 128 - temp_v0_12,
-                prim->b0 = temp_v0_12 + 192;
-                *((s32*)(&prim->r1)) = *((s32*)(&prim->r0));
-                *((s32*)(&prim->r2)) = *((s32*)(&prim->r0));
-                *((s32*)(&prim->r3)) = *((s32*)(&prim->r0));
-            }
-            if (self->accelerationY < 0) {
-                newEntity = AllocEntity(D_8007DE38, &D_8007DE38[24]);
-                if (newEntity != 0) {
-                    CreateEntityFromEntity(0x38, self, newEntity);
-                    newEntity->posX.i.hi -= 6 - ((Random() & 3) * 4);
-                    newEntity->zPriority = self->zPriority + 1;
-                }
-            }
-            if (self->accelerationY > 0) {
-                func_801C6458(0x1B);
-            }
-            if (self->ext.merman2.isUnderwater == 0) {
-                if (func_801BC8E4(&D_80182248) & 1) {
-                    g_api.FreePrimitives(self->firstPolygonIndex);
-                    self->hitboxHeight = 21;
-                    self->flags &= ~0x00800000;
-                    self->unk19 &= 0xFB;
-                    SetStep(5);
-                    return;
-                }
-            } else {
-                self->flags |= 0xC0000000;
-                MoveEntity();
-                self->accelerationY += 0x4000;
-            }
-        }
-
-        break;
-
-    case 5:
-        switch (self->step_s) {
-        case 0:
-            if (AnimateEntity(D_80182294, self) == 0) {
-                self->animFrameIdx = 0;
-                self->animFrameDuration = 0;
-                self->step_s++;
-            }
-            break;
-
-        case 1:
-            self->facing = (GetPlayerSide() & 1) ^ 1;
-            self->ext.merman2.timer = D_80182244[Random() & 3];
-            self->step_s++;
-            break;
-
-        case 2:
-            AnimateEntity(D_80182270, self);
-            colRes = func_801BCB5C(&D_80182268);
-            if (colRes == 0xFF) {
-                self->facing ^= 1;
-            }
-            if (self->facing == 0) {
-                self->accelerationX = -0x6000;
-            } else {
-                self->accelerationX = 0x6000;
-            }
-            if (colRes & 0x80) {
-                self->posX.val += self->accelerationX;
-            }
-            if ((!(self->subId & 1)) ||
-                ((self->accelerationX *= 2, !(self->subId & 1)))) {
-                if (!(self->posX.i.hi & 0xFF00)) {
-                    self->ext.merman2.timer--;
-                    if ((self->ext.merman2.timer & 0xFF) == 0) {
-                        if (Random() % 2) {
-                            SetStep(6);
-                        } else {
-                            self->step_s--;
-                        }
-                    }
-                    res = func_801C6458(0x1B);
-                    if (res != 0) {
-                        self->ext.merman2.ignoreCol = 1;
-                    }
-                }
-            } else {
-                res = func_801C6458(0x1B);
-                if (res != 0) {
-                    self->ext.merman2.ignoreCol = 1;
-                }
-            }
-        }
-
-        break;
-
-    case 6:
-        switch (self->step_s) {
-        case 0:
-            self->facing = (GetPlayerSide() & 1) ^ 1;
-            self->step_s++;
-            return;
-
-        case 1:
-            if (AnimateEntity(D_8018227C, self) == 0) {
-                func_801C2598(0x662);
-                newEntity = AllocEntity(D_8007A958, &D_8007A958[32]);
-                i = 0;
-                if (newEntity != 0) {
-                    CreateEntityFromEntity(0x34, self, newEntity);
-                    newEntity->posY.i.hi -= 12;
-                    newEntity->facing = self->facing;
-                }
-                temp_s4 = &D_8007A958[64];
-                for (offset = 0; i < 3; i++, offset += 8) {
-                    newEntity = AllocEntity(temp_s4, &temp_s4[32]);
-                    if (newEntity != 0) {
-                        CreateEntityFromEntity(0x36, self, newEntity);
-                        if (self->facing != 0) {
-                            newEntity->posX.i.hi += 8 + offset;
-                        } else {
-                            newEntity->posX.i.hi += (-8) - offset;
-                        }
-                        newEntity->posY.i.hi -= 12;
-                        newEntity->facing = self->facing;
-                        if (i == 0) {
-                            newEntity->subId = 1;
-                        }
-                    }
-                }
-
-                self->ext.merman2.rotation = 1;
-                self->unk1E = 0;
-                self->unk19 = self->unk19 | 4;
-                if (self->facing != 0) {
-                    self->accelerationX = -0x60000;
-                } else {
-                    self->accelerationX = 0x60000;
-                }
-                self->accelerationY = -0x1000;
-                self->posY.i.hi -= 1;
-                self->step_s++;
-            }
-            break;
-
-        case 2:
-            if ((self->ext.merman2.rotation < 0x100) && (g_blinkTimer & 1)) {
-                self->ext.merman2.rotation *= 2;
-            }
-            if (self->accelerationX != 0) {
-                if (self->facing != 0) {
-                    self->accelerationX += 0x3000;
-                } else {
-                    self->accelerationX -= 0x3000;
-                }
-            }
-            posX = self->posX.i.hi;
-            posY = self->posY.i.hi + 8;
-            if (self->facing != 0) {
-                posX -= 20;
-            } else {
-                posX += 20;
-            }
-            g_api.CheckCollision(posX, posY, &collider, 0);
-            if (collider.unk0 & 1) {
-                self->accelerationX = 0;
-            }
-            self->unk1E += self->ext.merman2.rotation;
-            self->accelerationY -= 0x2000;
-            if (func_801BC8E4(&D_80182248) & 1) {
-                if (self->facing == 0) {
-                    self->accelerationX = 0x28000;
-                } else {
-                    self->accelerationX = ~0x27FFF;
-                }
-                self->ext.merman2.rotation = 4;
-                self->posY.i.hi += 10;
-                self->step_s++;
-            }
-            func_801C6458(0x1B);
-            res = self->ext.merman2.isUnderwater;
-            if (res != 0) {
-                self->ext.merman2.ignoreCol = 1;
-            }
-            break;
-
-        case 3:
-            posY = self->posY.i.hi + 8;
-            posX = self->posX.i.hi;
-            if (self->facing != 0) {
-                posX -= 20;
-            } else {
-                posX += 20;
-            }
-            g_api.CheckCollision(posX, posY, &collider, 0);
-            if (collider.unk0 & 1) {
-                self->accelerationX = 0;
-            }
-            MoveEntity();
-            self->unk1E += 0xC0;
-            if (self->unk1E > 0x1000) {
-                self->posY.i.hi -= 10;
-                self->unk19 &= 0xFB;
-                SetStep(5);
-            }
-            res = func_801C6458(0x1B);
-            if (res != 0) {
-                self->ext.merman2.ignoreCol = 1;
-            }
-        }
-
-        break;
-
-    case 7:
-        if (self->step_s == 0) {
-            self->flags |= 0xC0000000;
-            self->step_s++;
-        }
-        MoveEntity();
-        self->accelerationY += 0x4000;
-        if ((!func_801C6458(0x1B)) && (!self->ext.merman2.isUnderwater)) {
-            self->ext.merman2.ignoreCol = 0;
-            SetStep(5);
-        }
-        break;
-
-    case 8:
-        switch (self->step_s) {
-        case 0:
-            self->animCurFrame = 14;
-            self->flags |= 0xC0000000;
-            if (!(GetPlayerSide() & 1)) {
-                self->accelerationX = -0x20000;
-            } else {
-                self->accelerationX = 0x20000;
-            }
-            self->accelerationY = -0x10000;
-            self->step_s++;
-            firstPrimIndex = g_api.AllocPrimitives(PRIM_GT4, 2);
-            if (firstPrimIndex != (-1)) {
-                prim = &g_PrimBuf[firstPrimIndex];
-                self->firstPolygonIndex = firstPrimIndex;
-                self->ext.merman2.prim = prim;
-                self->flags |= 0x800000;
-                func_801D2684(prim, firstPrimIndex);
-                prim->tpage = 0x12;
-                prim->clut = 0x292;
-                if (self->subId & 1) {
-                    prim->clut = 0x293;
-                }
-                if (self->facing != 0) {
-                    var_v0_11 = 0x1C8;
-                    prim->u0 = 0xF0;
-                    prim->u1 = var_v0_11;
-                } else {
-                    prim->u0 = 0xC8;
-                    prim->u1 = 0xF0;
-                }
-                prim->v0 = (prim->v1 = 0);
-                temp = prim->v0;
-                prim->u2 = prim->u0;
-                prim->u3 = prim->u1;
-                temp = prim->v0 + 0x30;
-                prim->v2 = temp;
-                prim->v3 = temp;
-                prim->priority = self->zPriority + 1;
-                prim->blendMode = 2;
-                *((s16*)(&prim->next->r2)) = 0x28;
-                *((s16*)(&prim->next->b2)) = 0x30;
-                prim->next->b3 = 0x80;
-                if (self->facing != 0) {
-                    prim->next->x1 = self->posX.i.hi - 3;
-                } else {
-                    prim->next->x1 = self->posX.i.hi + 3;
-                }
-                prim->next->y0 = self->posY.i.hi + 2;
-            }
-            func_801D1F38(prim);
-            break;
-
-        case 1:
-            func_801BC8E4(&D_80182258);
-            prim = *((s32*)(&self->ext.merman2.prim));
-            self->accelerationY -= 0x3000;
-            if (self->facing != 0) {
-                prim->next->x1 = self->posX.i.hi - 3;
-            } else {
-                prim->next->x1 = self->posX.i.hi + 3;
-            }
-            prim->next->y0 = self->posY.i.hi + 2;
-            func_801D1F38(prim);
-            if (!(g_blinkTimer % 10)) {
-                prim->clut += 2;
-                if (prim->clut >= 0x2A0) {
-                    prim->blendMode = 8;
-                    self->step_s++;
-                } else {
-                    newEntity = AllocEntity(D_8007D858, &D_8007D858[32]);
-                    if (newEntity != 0) {
-                        CreateEntityFromEntity(0x37, self, newEntity);
-                        newEntity->facing = self->facing;
-                        newEntity->subId = prim->clut;
-                        newEntity->zPriority = self->zPriority;
-                    }
-                }
-            }
-            func_801C6458(0x1B);
-            return;
-
-        case 2:
-            func_801C2598(0x65B);
-            func_801B653C();
-            DestroyEntity(self);
-        }
-
-        break;
-    }
-}
-#endif
-
-// some sort of explosion: ID 0x36
-void EntityExplosion2(Entity* entity, s32 arg1) {
-    Primitive* poly;
-    s16 firstPolygonIndex;
-
-    if (entity->step == 0) {
-        InitializeEntity(D_80180AB4);
-        entity->animCurFrame = 0;
-        entity->unk3C = 0;
-        entity->zPriority += 4;
-        if (entity->subId != 0) {
-            firstPolygonIndex = g_api.AllocPrimitives(PRIM_GT4, 2);
-            if (firstPolygonIndex == -1) {
-                DestroyEntity(entity);
-                return;
-            }
-            poly = &g_PrimBuf[firstPolygonIndex];
-            entity->firstPolygonIndex = firstPolygonIndex;
-            *(s32*)&entity->ext.generic.unk7C.s = poly;
-            entity->flags |= FLAG_FREE_POLYGONS;
-            func_801D2684(poly, firstPolygonIndex);
-            poly->u0 = 0;
-            poly->u1 = 0x20;
-            poly->tpage = 0x1A;
-            poly->clut = 0x1FF;
-            poly->v3 = poly->v2 = 0x20;
-            poly->v1 = poly->v0 = 0;
-            poly->u2 = poly->u0;
-            poly->u3 = poly->u1;
-            LOH(poly->next->r2) = 0x40;
-            LOH(poly->next->b2) = 0x40;
-            LOH(poly->next->u1) = 0;
-            poly->next->b3 = 0x60;
-            poly->next->x1 = entity->posX.i.hi;
-            poly->next->y0 = entity->posY.i.hi;
-            poly->priority = entity->zPriority - 4;
-            poly->blendMode = 6;
-        }
-    }
-
-    if (entity->subId != 0) {
-        poly = *(s32*)&entity->ext.generic.unk7C.s;
-        func_801D1F38(poly);
-        ((POLY_GT4*)poly->next)->b3 += 252;
-        LOH(poly->next->u1) -= 128;
-        if (poly->next->b3 < 16) {
-            poly->blendMode = 8;
-        }
-    }
-
-    entity->ext.generic.unk84.U8.unk0++;
-    if (!(entity->ext.generic.unk84.U8.unk0 % 4)) {
-        entity->posY.i.hi++;
-    }
-
-    if (AnimateEntity(D_801822B0, entity) == 0) {
-        DestroyEntity(entity);
-    }
-}
-
-void EntityMediumWaterSplash(Entity* entity) {
-    Entity* newEntity;
-
-    if (entity->step == 0) {
-        InitializeEntity(D_80180AC0);
-        entity->animCurFrame = 0;
-        if (entity->facing != 0) {
-            entity->accelerationX = 0x20000;
-            return;
-        }
-        entity->accelerationX = -0x20000;
-        return;
-    }
-    AnimateEntity(D_801822A4, entity);
-    MoveEntity();
-    if (entity->flags & 0x100) {
-        newEntity = AllocEntity(D_8007D858, &D_8007D858[MaxEntityCount]);
-        if (newEntity != NULL) {
-            CreateEntityFromEntity(ENTITY_EXPLOSION, entity, newEntity);
-            newEntity->subId = 0;
-        }
-        DestroyEntity(entity);
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3DEF4", func_801C7954);
-
-void func_801C7D80(Entity* self) {
-    if (self->step == 0) {
-        InitializeEntity(D_80180AB4);
-        self->animCurFrame = 0;
-        self->unk3C = 0;
-        self->flags |= 0x2000;
-        self->zPriority += 4;
-    }
-    MoveEntity();
-    self->accelerationY += 0x2800;
-    if (AnimateEntity(D_801822B8, self) == 0) {
-        DestroyEntity(self);
-    }
-}
-
-// ID 0x35
-void func_801C7E18(Entity* self) {
-    s8 temp; // probably !FAKE
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180A54);
-        self->animSet = 2;
-        self->accelerationY = -0x50000;
-        self->palette = 0x8162;
-        self->blendMode = 0x10;
-        self->palette = 0x8018;
-        self->blendMode = 0x30;
-        self->unk6C = 0xA0;
-        self->unk1A = 0x100;
-        self->unk1C = 0x1A0;
-        self->unk19 |= 3;
-        self->ext.generic.unk84.S8.unk1 = 0x11;
-        self->ext.generic.unk84.S8.unk0 = self->subId;
-        self->unk19 |= 8;
-        break;
-
-    case 1:
-        temp = self->ext.generic.unk84.S8.unk0 - 1;
-        self->ext.generic.unk84.S8.unk0--;
-        self->ext.generic.unk84.S8.unk1--;
-        if (!(temp & 0xFF)) {
-            self->step++;
-        }
-        break;
-
-    case 2:
-        AnimateEntity(D_801822C0, self);
-        MoveEntity();
-        self->accelerationY += 0x4000;
-        self->unk1A += 6;
-        self->unk1C -= 4;
-        if (self->posY.i.hi > 256) {
-            DestroyEntity(self);
-        }
-        if (!--self->ext.generic.unk84.U8.unk1) {
-            self->accelerationY = 0;
-            self->step++;
-        }
-        break;
-
-    case 3:
-        if (AnimateEntity(D_801822C0, self) == 0) {
-            MoveEntity();
-            self->accelerationY += 0x4000;
-            self->unk1A += 6;
-            self->unk1C -= 4;
-        }
-        if (self->posY.i.hi > 256) {
-            DestroyEntity(self);
-        }
-        break;
-    }
-}
